code organization, progress
